Refactor: Unify components and styles
- I removed wrapper components for LanguageSelector and used CosyLanguageSelector directly in FreestyleModePage and StudyModePage.
- I created a shared CSS file and moved common styles to it.
