+++ src/pages/FreestyleModePage/FreestyleModePage.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { useFreestyle } from '../../contexts/FreestyleContext';
-import LanguageSelectorFreestyle from '../../components/Freestyle/LanguageSelectorFreestyle';
+import { useI18n } from '../../i18n/I18nContext';
+import CosyLanguageSelector from '../../components/LanguageSelector/CosyLanguageSelector';
 import DaySelectorFreestyle from '../../components/Freestyle/DaySelectorFreestyle';
 import PracticeCategoryNav from '../../components/Freestyle/PracticeCategoryNav';
 import ExerciseHost from '../../components/Freestyle/ExerciseHost';
@@ -13,7 +14,13 @@ import '../../freestyle-shared.css';
 import './FreestyleModePage.css';
 
 const FreestyleModePage = () => {
-  const { selectedLanguage, selectedDays, selectedExercise } = useFreestyle();
+  const { selectedLanguage, selectedDays, selectedExercise, setSelectedLanguage } = useFreestyle();
+  const { language, changeLanguage } = useI18n();
+
+  const handleLanguageChange = (newLanguage) => {
+    changeLanguage(newLanguage);
+    setSelectedLanguage(newLanguage);
+  };
 
   const words = [
     { text: 'un', size: 1.5 },
@@ -37,7 +44,10 @@ const FreestyleModePage = () => {
       <FreestyleProgress />
 
       <div className="freestyle-controls-container">
-        <LanguageSelectorFreestyle />
+        <CosyLanguageSelector
+          selectedLanguage={language}
+          onLanguageChange={handleLanguageChange}
+        />
         {selectedLanguage && <DaySelectorFreestyle language={selectedLanguage} />}
         {selectedLanguage && selectedDays.length > 0 && (
           <PracticeCategoryNav language={selectedLanguage} days={selectedDays} />
