+++ src/styles/shared.css
@@ -1,0 +1,27 @@
+.page-container {
+    padding: 20px;
+    background-color: #f0f8ff;
+}
+
+.page-header {
+    text-align: center;
+    margin-bottom: 20px;
+    color: #333;
+}
+
+.island-placeholder {
+    background-color: #ffffff;
+    border-radius: 12px;
+    padding: 20px;
+    margin-bottom: 20px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.05);
+    border: none;
+}
+
+.island-placeholder {
+  padding: 15px;
+  margin-bottom: 15px;
+  border: 1px dashed #ccc; /* Consider var(--color-border) */
+  border-radius: 4px; /* Consider var(--border-radius-md) */
+  background-color: #f9f9f9; /* Consider var(--color-surface) */
+}
